clean up main: move methods call and create game loop example
move startup checks from main to Detections

call startup checks and set `UnmanagedGlobals::SupressingNewThreads` in AntiCheat constructor.

added AntiCheatInitFail errors for the constructor, for the startup checks.

create Constructor for AntiCheat that automatically calls `GetWindowsVersion`

move `UnmanagedGlobals` methods in Globals.cpp

created Init.cpp/hpp to setup the `TLSCallback`

cleaned some parts of the main, moved calls to settings/banners printing to a method at the bottom of main to unclutter the main, as it is the first code read by new users.

created small game loop example in main , that shows new detections, and stops the program when any key is pressed.

todo: move Thread tampering detection to its own feature instead of putting it in the main.
