+++ Common/Globals.cpp
@@ -1,0 +1,63 @@
+#include "Globals.hpp"
+
+/*
+    AddThread - adds a Thread* object to our global thread list
+*/
+bool UnmanagedGlobals::AddThread(DWORD id)
+{
+    DWORD tid = GetCurrentThreadId();
+    Logger::logf(Info, " New thread spawned: %d", tid);
+
+    CONTEXT context;
+    context.ContextFlags = CONTEXT_ALL;
+
+    HANDLE threadHandle = OpenThread(THREAD_ALL_ACCESS, FALSE, tid);
+
+    if (threadHandle == NULL)
+    {
+        Logger::logf(Warning, " Couldn't open thread handle @ TLS Callback: Thread %d", tid);
+        return false;
+    }
+    else
+    {
+        Thread* t = new Thread(tid); //memory must be free'd after done using, this function does not free mem
+        UnmanagedGlobals::ThreadList->push_back(t);
+        return true;
+    }
+}
+
+/*
+    RemoveThread - Removes Thread* with threadid `tid` from our global thread list
+*/
+void UnmanagedGlobals::RemoveThread(DWORD tid)
+{
+    Thread* ToRemove = NULL;
+
+    list<Thread*>::iterator it;
+
+    for (it = ThreadList->begin(); it != ThreadList->end(); ++it)
+    {
+        Thread* t = it._Ptr->_Myval;
+        if (t->GetId() == tid)
+            ToRemove = t;
+    }
+
+    if (ToRemove != NULL) //remove thread from our list on thread_detach
+        ThreadList->remove(ToRemove);
+}
+
+/*
+    ExceptionHandler - User defined exception handler which catches program-wide exceptions
+    ...Currently we are not doing anything special with this, but we'll leave it here incase we need it later
+*/
+LONG WINAPI UnmanagedGlobals::ExceptionHandler(EXCEPTION_POINTERS* ExceptionInfo)  //handler that will be called whenever an unhandled exception occurs in any thread of the process
+{
+    DWORD exceptionCode = ExceptionInfo->ExceptionRecord->ExceptionCode;
+    Logger::logf(Warning, "Program threw exception: %x at %llX\n", exceptionCode, ExceptionInfo->ExceptionRecord->ExceptionAddress);
+    
+    if (exceptionCode == EXCEPTION_BREAKPOINT) //one or two of our debug checks may throw this exception
+    {
+    } //optionally we may be able to view the exception address and compare it to whitelisted module address space, if it's not contained then we assume it's attacker-run code
+
+    return EXCEPTION_CONTINUE_SEARCH;
+}
