+++ Init.hpp
@@ -1,0 +1,20 @@
+#include "API/API.hpp"
+
+#pragma comment(linker, "/ALIGN:0x10000") //for remapping technique (anti-tamper) - each section gets its own region, align with system allocation granularity
+
+using namespace std;
+
+void NTAPI __stdcall TLSCallback(PVOID pHandle, DWORD dwReason, PVOID Reserved);
+
+#pragma comment (linker, "/INCLUDE:_tls_used")
+#pragma comment (linker, "/INCLUDE:_tls_callback")
+
+EXTERN_C
+#ifdef _M_X64
+#pragma const_seg (".CRT$XLB")
+const
+#endif
+
+PIMAGE_TLS_CALLBACK _tls_callback = TLSCallback;
+#pragma data_seg ()
+#pragma const_seg ()
